// app/components/theme-provider/theme/types.ts
export type TBackground =
  | "color.background.body"
  | "color.background.component";

export type TLink = "color.link" | "color.link.hover" | "color.link.active";
// ...
export type TThemeConfigurationVariableKey =
  | "color.white"
  | "color.blue"
  | "color.green"
  | "color.cyan"
  | "color.yellow"
  | "color.orange"
  | "color.red"
  | "color.purple"
  | "color.pink"
  | "color.primary"
  | "color.success"
  | "color.info"
  | "color.warning"
  | "color.error"
  | TBackground
  | TLink
  | "spacing"
  | "font.family"
  | "font.size"
  | "font.size.small"
  | "font.size.large"
  | "font.color"
  | "line.height"
  | "border.color"
  | "border.radius"
  | "border.radius.small"
  | "border.radius.large";

export type TThemeConfigurationVariables = Partial<
  Record<TThemeConfigurationVariableKey, string>
>;